pyproject.toml
[tool.poetry]
name = "Execution Engine Test"
version = "0.0.1"
description = ""
authors = ["Ira Rainey"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
docker = "^7.1.0"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
azure-identity = "^1.19.0"
kubernetes = "^32.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
flake8 = "^7.1.1"
pylint = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
